Create qrcodesss
The code creates a QR code for a given piece of data (in this case, a URL) and saves it as an image file. Here’s a breakdown of what the code does:

Imports the qrcode library:

This library is used to generate QR codes.
Defines the function create_qr_code(data, filename):

Creates an instance of a QR code:

version=1: Controls the size of the QR code. A lower number means a smaller QR code.
error_correction=qrcode.constants.ERROR_CORRECT_L: Specifies the error correction level. L allows for about 7% error correction.
box_size=10: Determines the number of pixels each box of the QR code will be.
border=4: Sets the thickness of the border (in boxes).
Adds data to the QR code:

qr.add_data(data): Adds the data (e.g., a URL) to be encoded in the QR code.
qr.make(fit=True): Ensures the QR code is the appropriate size to fit the data.
Creates an image from the QR code instance:

img = qr.make_image(fill='black', back_color='white'): Creates an image with the QR code, using black for the QR code and white for the background.
Saves the image to a file:

img.save(filename): Saves the generated QR code image to the specified filename.
Main script execution:

Sets the data to be encoded in the QR code (a URL in this case) and the filename for the output image.
Calls the create_qr_code function with the data and filename.
Prints a confirmation message that the QR code has been saved.